backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
redis==5.0.4
fakeredis==2.23.2
chromadb==0.5.3
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
langgraph==0.1.5
langchain==0.2.6
langchain-community==0.2.6
anthropic==0.28.0
openai==1.35.0
groq==0.9.0
requests==2.32.3
pyngrok==7.2.0
